+++ tests/array.global.test.mjs
@@ -43,4 +43,5 @@ test('array.global', () => {
     expect(array_walk).not.toBeUndefined();
     expect(array_walk_recursive).not.toBeUndefined();
     expect(in_array).not.toBeUndefined();
+    expect(key_exists).not.toBeUndefined();
 });